requirements.txt
# 基础依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI相关
openai>=1.0.0
langchain>=0.0.300
tiktoken>=0.5.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 自然语言处理
nltk>=3.8.0
jieba>=0.42.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志和监控
loguru>=0.7.0
schedule>=1.2.0

# 网络工具
aiohttp>=3.8.0
httpx>=0.24.0

# 工具类
tqdm>=4.65.0
colorama>=0.4.0